[Bronze II] Title: Hashing, Time: 124 ms, Memory: 14260 KB -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -7,15 +7,15 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
         String str = br.readLine();
-        int sum = 0;
-        
+        double sum = 0;
+
         for (int i = 0; i < N; i++) {
             char c = str.charAt(i);
             int num = c - 96;
             sum += num * Math.pow(31, i);
         }
 
-        System.out.println(sum);
+        System.out.println((long)sum);
 
     }
 }
